DB cleanup fix
