+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/ha/SlickDistributedLockSpec.scala
@@ -7,13 +7,21 @@ import pl.touk.nussknacker.test.base.db.WithPostgresDbTesting
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
-import scala.util.Using
+import scala.util.{Try, Using}
 
 @Slow
 class SlickDistributedLockSpec extends AnyFunSuite with Matchers with WithPostgresDbTesting {
 
   private val lockName = "test-lock"
   private val duration = 30.seconds
+
+  override def cleanDB(): Try[Unit] = {
+    val superResult = super.cleanDB()
+    val locksResult = Using(testDbRef.db.createSession()) { session =>
+      session.prepareStatement(s"""DELETE FROM "${getSchemaName()}"."distributed_locks"""").execute()
+    }.map(_ => ())
+    superResult.flatMap(_ => locksResult)
+  }
 
   private def lock(instanceId: String) = SlickDistributedLock(testDbRef, instanceId, lockQueryTimeout = 5.seconds)
   private lazy val inst1               = lock("instance-1")
